database implementation added

+++ src/process_data.rs
@@ -3,8 +3,8 @@ pub struct ProcessData {
     pub pid : u32,
     pub ppid : u32,
     pub process_name : String,
-    pub cpu_usage : u32,
-    pub ram_usage : u32,
+    pub cpu_usage : f64,
+    pub ram_usage : f64,
     pub state : String,
     pub threads_used : u32,
 }
